QA-15375: Fix localized session resolution, add localized property mutation
